[FEATURE] Add capsule colliders and separate player visuals
Implement Y-aligned capsule collision, serialization, ray detection, and dedicated debug meshes with validation and coverage. Keep imported player renderables on a child visual object so gameplay transforms and capsule geometry remain independent, and document the longer-term scene-node refactor.

+++ src/resource_loader/resource_loader.cpp
@@ -102,7 +102,12 @@ Maths::Transform node_transform(const tinygltf::Node& node)
 	if (!node.translation.empty())
 		transform.set_pos({ node.translation[0], node.translation[1], node.translation[2] });
 	if (!node.rotation.empty())
-		transform.set_orient({ node.rotation[3], node.rotation[0], node.rotation[1], node.rotation[2] });
+		transform.set_orient({
+			static_cast<float>(node.rotation[3]),
+			static_cast<float>(node.rotation[0]),
+			static_cast<float>(node.rotation[1]),
+			static_cast<float>(node.rotation[2])
+		});
 	if (!node.scale.empty())
 		transform.set_scale({ node.scale[0], node.scale[1], node.scale[2] });
 	return transform;
